fix: incorrect executable metadata on toolchain versions >= 1.61
Fixes an issue where building xterminate on Rust toolchains version 1.61 and newer causes the winres crate to silently fail to embed the manifest and rc files into the executable, resulting in default ones being embedded instead.

The issue is fixed by replacing the non-maintained winres crate with winresource -- a maintained fork of winres with a fix for the issue.